Refactor out common pnp quota abstract class
wip start refactor - tests pass

wip checkpoint: factored out the common pnp quota class

wip checkpoint pre:rename

wip remove test comment
