+++ packages/phone-number-privacy/signer/src/pnp/services/quota.legacy.ts
@@ -1,5 +1,5 @@
 import { NULL_ADDRESS } from '@celo/base'
-import { ContractKit, StableToken } from '@celo/contractkit'
+import { StableToken } from '@celo/contractkit'
 import {
   ErrorMessage,
   isVerified,
@@ -7,13 +7,9 @@ import {
   SignMessageRequest,
 } from '@celo/phone-number-privacy-common'
 import Logger from 'bunyan'
-import { Knex } from 'knex'
-import { getPerformedQueryCount, incrementQueryCount } from '../../common/database/wrappers/account'
-import { storeRequest } from '../../common/database/wrappers/request'
 import { Counters, Histograms } from '../../common/metrics'
-import { OdisQuotaStatusResult, QuotaService } from '../../common/quota'
+import { QuotaService } from '../../common/quota'
 import {
-  getBlockNumber,
   getCeloBalance,
   getStableTokenBalance,
   getTransactionCount,
@@ -22,110 +18,17 @@ import {
 } from '../../common/web3/contracts'
 import { config } from '../../config'
 import { PnpSession } from '../session'
+import { PnpQuotaServiceCommonRename } from './quota_common_rename'
 export interface PnpQuotaStatus {
   queryCount: number
   totalQuota: number
   blockNumber: number
 }
 
-export class LegacyPnpQuotaService implements QuotaService<SignMessageRequest | PnpQuotaRequest> {
-  constructor(readonly db: Knex, readonly kit: ContractKit) {}
-
-  public async checkAndUpdateQuotaStatus(
-    state: PnpQuotaStatus,
-    session: PnpSession<SignMessageRequest>,
-    trx: Knex.Transaction
-  ): Promise<OdisQuotaStatusResult<SignMessageRequest>> {
-    const remainingQuota = state.totalQuota - state.queryCount
-    Histograms.userRemainingQuotaAtRequest.observe(remainingQuota)
-    let sufficient = remainingQuota > 0
-    if (!sufficient) {
-      session.logger.debug({ ...state }, 'No remaining quota')
-      if (this.bypassQuotaForE2ETesting(session.request.body)) {
-        Counters.testQuotaBypassedRequests.inc()
-        session.logger.info(
-          { request: session.request.body },
-          'Request will bypass quota check for e2e testing'
-        )
-        sufficient = true
-      }
-    } else {
-      await this.updateQuotaStatus(trx, session)
-      state.queryCount++
-    }
-    return { sufficient, state }
-  }
-
-  public async getQuotaStatus(
-    session: PnpSession<SignMessageRequest | PnpQuotaRequest>,
-    trx?: Knex.Transaction
-  ): Promise<PnpQuotaStatus> {
-    const { account } = session.request.body
-    const [queryCountResult, totalQuotaResult, blockNumberResult] = await meter(
-      (_session: PnpSession<SignMessageRequest | PnpQuotaRequest>) =>
-        Promise.allSettled([
-          // TODO(Alec)(pnp)
-          // Note: The database read of the user's queryCount
-          // included here resolves to 0 on error
-          getPerformedQueryCount(this.db, account, session.logger, trx),
-          this.getTotalQuota(_session),
-          getBlockNumber(this.kit),
-        ]),
-      [session],
-      (err: any) => {
-        throw err
-      },
-      Histograms.getRemainingQueryCountInstrumentation,
-      'getQuotaStatus'
-    )
-
-    let hadBlockchainError = false
-    let queryCount = -1
-    let totalQuota = -1
-    let blockNumber = -1
-    if (queryCountResult.status === 'fulfilled') {
-      queryCount = queryCountResult.value
-    } else {
-      session.logger.error(queryCountResult.reason)
-      session.errors.push(ErrorMessage.DATABASE_GET_FAILURE)
-    }
-    if (totalQuotaResult.status === 'fulfilled') {
-      totalQuota = totalQuotaResult.value
-    } else {
-      session.logger.error(totalQuotaResult.reason)
-      hadBlockchainError = true
-    }
-    if (blockNumberResult.status === 'fulfilled') {
-      blockNumber = blockNumberResult.value
-    } else {
-      session.logger.error(blockNumberResult.reason)
-      hadBlockchainError = true
-    }
-    if (hadBlockchainError) {
-      session.errors.push(ErrorMessage.CONTRACT_GET_FAILURE)
-    }
-
-    return { queryCount, totalQuota, blockNumber }
-  }
-
-  protected async updateQuotaStatus(
-    trx: Knex.Transaction,
-    session: PnpSession<SignMessageRequest>
-  ) {
-    // TODO(Alec)(pnp): Review db error handling
-    const [requestStored, queryCountIncremented] = await Promise.all([
-      storeRequest(this.db, session.request.body, session.logger, trx),
-      incrementQueryCount(this.db, session.request.body.account, session.logger, trx),
-    ])
-    if (!requestStored) {
-      session.logger.debug('Did not store request.')
-      session.errors.push(ErrorMessage.FAILURE_TO_STORE_REQUEST)
-    }
-    if (!queryCountIncremented) {
-      session.logger.debug('Did not increment query count.')
-      session.errors.push(ErrorMessage.FAILURE_TO_INCREMENT_QUERY_COUNT)
-    }
-  }
+export class LegacyPnpQuotaService
+  extends PnpQuotaServiceCommonRename
+  implements QuotaService<SignMessageRequest | PnpQuotaRequest> {
+  protected readonly metricsPrefix = 'LegacyPnpQuotaService.'
 
   protected async getWalletAddressAndIsVerified(
     session: PnpSession<SignMessageRequest | PnpQuotaRequest>
@@ -225,26 +128,12 @@ export class LegacyPnpQuotaService implements QuotaService<SignMessageRequest | 
     return { cUSDAccountBalance, cEURAccountBalance, celoAccountBalance }
   }
 
-  protected async getTotalQuota(
-    session: PnpSession<SignMessageRequest | PnpQuotaRequest>
-  ): Promise<number> {
-    return meter(
-      this.getTotalQuotaWithoutMeter,
-      [session],
-      (err: any) => {
-        throw err
-      },
-      Histograms.getRemainingQueryCountInstrumentation,
-      'getTotalQuota'
-    )
-  }
-
   /*
    * Calculates how many queries the caller has unlocked based on the algorithm
    * unverifiedQueryCount + verifiedQueryCount + (queryPerTransaction * transactionCount)
    * If the caller is not verified, they must have a minimum balance to get the unverifiedQueryMax.
    */
-  private async getTotalQuotaWithoutMeter(
+  protected async getTotalQuotaWithoutMeter(
     session: PnpSession<SignMessageRequest | PnpQuotaRequest>
   ): Promise<number> {
     const {
@@ -381,10 +270,5 @@ export class LegacyPnpQuotaService implements QuotaService<SignMessageRequest | 
     })
 
     return quota
-  }
-
-  private bypassQuotaForE2ETesting(requestBody: SignMessageRequest): boolean {
-    const sessionID = Number(requestBody.sessionID)
-    return !Number.isNaN(sessionID) && sessionID % 100 < config.test_quota_bypass_percentage
   }
 }
